agent-evolution/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
claude-agent-sdk
python-dotenv>=1.0.0
pydantic>=2.0.0
